wp7: run application without models

+++ lib/framework/rho/rho.rb
@@ -190,6 +190,8 @@ if defined?( RHODES_EMULATOR )
     end
 else
     def load_models_from_file(app_manifest_filename=nil)
+		return unless Rho::file_exist?(app_manifest_filename)
+
         f = File.open(app_manifest_filename)
         _load_models_from_file(f)    
         f.close
